# packages/purebackend/docs/swagger.yaml
definitions:
...
  models.HashRequest:
    properties:
      hash:
        type: string
    type: object
...
info:
  contact:
    email: contact@pureml.com
    name: API Support
    url: http://www.swagger.io/support
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
paths:
...
  /org/{orgId}/model/{modelName}/branch/{branchName}/hash-status:
    post:
      consumes:
      - '*/*'
      description: Verify model hash status to determine if model is already uploaded
      parameters:
      - description: Organization Id
        in: path
        name: orgId
        required: true
        type: string
      - description: Model Name
        in: path
        name: modelName
        required: true
        type: string
      - description: Branch Name
        in: path
        name: branchName
        required: true
        type: string
      - description: Hash value
        in: body
        name: hash
        required: true
        schema:
          $ref: '#/definitions/models.HashRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Verify model hash status
      tags:
      - Model
  /org/{orgId}/model/{modelName}/branch/{branchName}/register:
    post:
      consumes:
      - '*/*'
      description: Register model file. Create model and default branches if not exists
      parameters:
      - description: Model file
        in: formData
        name: file
        required: true
        type: file
      - description: Organization UUID
        in: path
        name: orgId
        required: true
        type: string
      - description: Model name
        in: path
        name: modelName
        required: true
        type: string
      - description: Branch name
        in: path
        name: branchName
        required: true
        type: string
      - in: formData
        name: hash
        type: string
      - in: formData
        name: is_empty
        type: boolean
      - in: formData
        name: storage
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Register model
      tags:
      - Model
...
  /org/{orgId}/model/{modelName}/branch/{branchName}/version/{version}:
    get:
      consumes:
      - '*/*'
      description: Get specific branch version of a model
      parameters:
      - description: Organization Id
        in: path
        name: orgId
        required: true
        type: string
      - description: Model Name
        in: path
        name: modelName
        required: true
        type: string
      - description: Branch Name
        in: path
        name: branchName
        required: true
        type: string
      - description: Version
        in: path
        name: version
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Get specific branch version of a model
      tags:
      - Model
...
securityDefinitions:
  ApiKeyAuth:
    description: 'Header for logged in user format: Bearer {token}'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"